Introduce notion of `Renderer`
A renderer takes a structure used in a dialect (e.g. JSON, RDF) and translates it into a string
