+++ src/metadata_tool/cli.py
@@ -34,7 +34,7 @@ def translate(f, t, o, file_path):
         from_dialect = get_dialect(f)()
         obj = from_dialect.parse(infile.read())
         to_dialect = get_dialect(t)()
-        s = to_dialect.compile(obj)
+        s = to_dialect.compile_and_render(obj)
         if o:
             with open(o, "w") as outfile:
                 outfile.write(s.decode("utf-8"))
